Aplha 0.5 Update 10
Triggers have been added to change the behavior of the player, and most of the tutorial objects have been fixed for a bug that caused them not to crouch correctly when climbing in a low-ceilinged area.

A bug causing the DodgeRoll to run in queue is being investigated, but the movement of the DodgeRoll has been fixed.